[RF] Silence member shadowing warning with gcc48
It's a problem of old gccs. New compilers consider naming a variable the
same as a function not a shadow. This was fixed in gcc5, although the
release notes of gcc 4.8 claim that it was in 4.8. That's not the case,
as the warning that popped up in the nightlies showed.

This commit silences these warnings by renaming the relevant variables.