[DF] Add RDataFrame::Redefine
This method behaves similarly to Define but it allows overriding
the lookup of an already existing column.

This is useful e.g. to write a new TTree with changed values for
a given branch (without being forced to pick a different name for
the new branch) or to turn on/off corrections and systematic
variations without changing downstream logic in the computation graph.

RedefineSlot and RedefineSlotEntry are also available, and passing
strings to be just-in-time compiled is also supported.