[DF] Avoid using GetBranch to check for column in TTree
Instead, check whether the column is in the one true list of branch
names kept by RLoopManager. This is a bit better than interrogating
the TTree directly because correct usage of GetBranch, FindBranch,
GetLeaf and FindLeaf can be tricky; so let's assume we got it right
when we collected the list of available branches -- or at least let's
be consistent in which branch names we consider valid.