[DF] Re-introduce warning for un-triggered lazy Snapshots
RDF used to warn in case a lazy Snapshot was booked but never triggered.
The logic that triggered the warning silently stopped working after
we switched to the generic jitting logic for Snapshot and we changed the
ownership model of RDataFrame's nodes in the computation graph.

We used to call SnapshotHelper[MT]::Finalize even if the event
loop was never triggered, but in general this is not the case anymore:
now we don't even create the SnapshotHelper[MT] helper object (in the
case of jitted Snapshots) or anyway might not call Finalize on it if the
event loop is not triggered.

With this patch we move the check for "booked but never triggered"
inside the destructor of RResultPtr<RInterface<RLoopManager>> instead,
which we can be sure will always be called.